style: moved interfaces
