+++ src/app/features/event/event-list/event-list.component.ts
@@ -4,8 +4,7 @@ import { Store } from '@ngrx/store';
 import { ErrorhandlerService } from '@shared/Interceptor/errorhandler.service';
 import { Subject, takeUntil } from 'rxjs';
 import { AppState } from 'src/app/app.module';
-import { FilterValue, SortValue } from '../event.interfaces';
-import { GetEventPayload } from './event-list.interface';
+import { FilterValue, GetEventPayload, SortValue } from './event-list.interface';
 import { EventListService } from './event-list.service';
 import { filterOptions, sortOptions } from './filterSortOptions';
 import { EventListActions } from './store/event-list.actions';
