+++ src/app/features/event/event-list/event-list.interface.ts
@@ -1,4 +1,4 @@
-import { EventCard, FilterValue, SortValue } from '../event.interfaces';
+import { EventCard } from '../event.interfaces';
 
 export interface EventListState {
   events: EventCard[];
@@ -9,4 +9,18 @@ export interface GetEventPayload {
   sort: SortValue;
   filter: FilterValue[];
   category: string | null;
+}
+
+export type SortValue = 'sortOrder=desc' | 'sortOrder=asc' | 'participants=asc' | 'participants=desc';
+
+export type FilterValue = 'participating=true' | 'group=true';
+
+export interface SortOption {
+  name: string;
+  value: SortValue;
+}
+
+export interface FilterOption {
+  name: string;
+  value: FilterValue;
 }
