+++ src/app/features/event/event.interfaces.ts
@@ -58,17 +58,3 @@ export interface EventSurvey {
     id?: number;
   }[];
 }
-
-export type SortValue = 'sortOrder=desc' | 'sortOrder=asc' | 'participants=asc' | 'participants=desc';
-
-export type FilterValue = 'participating=true' | 'group=true';
-
-export interface SortOption {
-  name: string;
-  value: SortValue;
-}
-
-export interface FilterOption {
-  name: string;
-  value: FilterValue;
-}
